Adds soft limits to armIOSparkMax. values should be tested

+++ src/main/java/frc/robot/constants/ArmConstants.java
@@ -17,8 +17,8 @@ public class ArmConstants {
     public static final double STARTING_POSITION = Units.degreesToRadians(-55);
 
     // TODO implement and tune the following :)
-    public static final double MAX_POSITION = 90; // Soft stop for max (farthest from floor) position
-    public static final double MIN_POSITION = 90; // Soft stop for min (closest to floor) position
+    public static final float MAX_POSITION = 90; // Soft stop for max (farthest from floor) position
+    public static final float MIN_POSITION = 90; // Soft stop for min (closest to floor) position
 
     // Current limit for arm motor.
     public static final int CURRENT_LIMIT = 15;
